+++ cmd/runner.go
@@ -1,9 +1,9 @@
 package cmd
 
 import (
+	"fmt"
 	"io"
 
-	"github.com/giantswarm/microerror"
 	"github.com/spf13/cobra"
 )
 
@@ -15,7 +15,7 @@ type runner struct {
 func (r *runner) Run(cmd *cobra.Command, args []string) error {
 	err := cmd.Help()
 	if err != nil {
-		return microerror.Mask(err)
+		return fmt.Errorf("failed to show help: %w", err)
 	}
 
 	return nil